third_party/blink/web_tests/external/wpt/css/css-values/calc-z-index-fractions-001-expected.txt

This is a testharness.js-based test.
FAIL testing z-index: calc(3 / 2) assert_equals: expected "2" but got "auto"
FAIL testing z-index: calc(-3 / 2) assert_equals: expected "-1" but got "auto"
Harness: the test ran to completion.
